+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Support for loading configration from custom URL [PR #323](https://github.com/3scale/apicast/pull/323)
 - Turn on SSL/TLS validation by `OPENSSL_VERIFY` environment variable [PR #332](https://github.com/3scale/apicast/pull/332)
 - Load trusted CA chain certificates [PR #332](https://github.com/3scale/apicast/pull/332)
+- Support HTTP Basic authentication for client credentials when authorizing with RH-SSO [PR #336](https://github.com/3scale/apicast/pull/336)
 
 ### Changed